+++ scripts/Muon/GUI/Common/plot_widget/plot_widget_model.py
@@ -6,6 +6,7 @@
 # SPDX - License - Identifier: GPL - 3.0 +
 from Muon.GUI.FrequencyDomainAnalysis.frequency_context import FREQUENCY_EXTENSIONS
 from Muon.GUI.Common.ADSHandler.workspace_naming import TF_ASYMMETRY_PREFIX
+from Muon.GUI.Common.contexts.frequency_domain_analysis_context import FrequencyDomainAnalysisContext
 
 COUNTS_PLOT_TYPE = 'Counts'
 ASYMMETRY_PLOT_TYPE = 'Asymmetry'
@@ -135,7 +136,7 @@ class PlotWidgetModel(object):
 
     def get_plot_types(self):
         plot_types = [ASYMMETRY_PLOT_TYPE, COUNTS_PLOT_TYPE]
-        if self.context._frequency_context:
+        if isinstance(self.context, FrequencyDomainAnalysisContext):
             for ext in FREQUENCY_EXTENSIONS.keys():
                 plot_types.append(FREQ_PLOT_TYPE + FREQUENCY_EXTENSIONS[ext])
             plot_types.append(FREQ_PLOT_TYPE + "All")
