+++ scripts/Muon/GUI/MuonAnalysis/muon_analysis_2.py
@@ -11,7 +11,7 @@ from distutils.version import LooseVersion
 from mantid.kernel import ConfigServiceImpl
 
 import Muon.GUI.Common.message_box as message_box
-from Muon.GUI.Common.contexts.muon_context import MuonContext
+from Muon.GUI.Common.contexts.data_analysis_context import DataAnalysisContext
 from Muon.GUI.Common.contexts.muon_data_context import MuonDataContext
 from Muon.GUI.Common.contexts.muon_group_pair_context import MuonGroupPairContext
 from Muon.GUI.Common.contexts.phase_table_context import PhaseTableContext
@@ -78,12 +78,11 @@ class MuonAnalysisGui(QtWidgets.QMainWindow):
         self.phase_context = PhaseTableContext()
         self.fitting_context = FittingContext()
 
-        self.context = MuonContext(muon_data_context=self.data_context,
+        self.context = DataAnalysisContext(muon_data_context=self.data_context,
                                    muon_gui_context=self.gui_context,
                                    muon_group_context=self.group_pair_context,
-                                   muon_phase_context=self.phase_context,
                                    fitting_context=self.fitting_context,
-                                   workspace_suffix=' MA')
+                                   muon_phase_context=self.phase_context)
 
         # create the Dockable plot widget
         self.plot_widget = PlotWidget(self.context, parent=self)
